# stario-platform/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Stario API
  description: |
    Stario / TabrikStar - AI Emotion Platform API

    Create personalized AI video greetings, face quizzes, posters, and more
    with your favorite artists.

    ## Authentication
    Most endpoints require Bearer token authentication.
    Get your token via `/auth/login` or `/auth/telegram`.

    ## Rate Limiting
    - 60 requests per minute per user
    - 10,000 requests per day per user

    ## Compliance (Uzbekistan 2025-2030)
    - Data residency: All user data stored within Uzbekistan
    - PII minimization: Face Quiz photos deleted within 3 seconds
    - Audit logs retained for 90 days

  version: 1.0.0
  contact:
    name: Stario Support
    email: support@stario.uz
servers:
  - url: https://api.stario.uz
    description: Production
  - url: https://api-staging.stario.uz
    description: Staging
  - url: http://localhost:8000
    description: Local development
...
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '409':
          description: Email already registered
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with email/password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
  /auth/telegram:
    post:
      tags: [Authentication]
      summary: Login via Telegram Mini App
      operationId: telegramAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramAuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
...
  /face-quiz/start:
    post:
      tags: [Face Quiz]
      summary: Start face quiz
      description: |
        Begin face similarity comparison with an artist.
        Returns presigned URL for direct photo upload.
        Photos are deleted within 3 seconds per PII compliance.
      operationId: startFaceQuiz
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaceQuizRequest'
      responses:
        '200':
          description: Quiz started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceQuizResponse'
  /face-quiz/{quizId}/analyze:
    post:
      tags: [Face Quiz]
      summary: Analyze face similarity
      description: |
        Process uploaded photo and return similarity score.
        Target processing time: ≤200ms.
      operationId: analyzeFace
      security:
        - bearerAuth: []
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceSimilarityResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
        phone:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TelegramAuthRequest:
      type: object
      required: [init_data]
      properties:
        init_data:
          type: string
          description: Telegram Mini App init data
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
...
    FaceQuizRequest:
      type: object
      required: [artist_id]
      properties:
        artist_id:
          type: string
        save_photo:
          type: boolean
          default: false
          description: If false, photo deleted after 3s per PII rules
    FaceQuizResponse:
      type: object
      properties:
        quiz_id:
          type: string
        upload_url:
          type: string
          description: Presigned URL for direct upload
        upload_key:
          type: string
        expires_in_seconds:
          type: integer
    FaceSimilarityResult:
      type: object
      properties:
        quiz_id:
          type: string
        artist_id:
          type: string
        similarity_score:
          type: number
          minimum: 0
          maximum: 100
        matching_features:
          type: array
          items:
            type: string
        rank_percentile:
          type: number
        badge_earned:
          type: string
        share_image_url:
          type: string
...